Ci integration
* Added github CI integration

* windows build

* release build

* fix return type

* fix some warnings

+++ ssh/crypto/random.cpp
@@ -2,6 +2,7 @@
 
 #include "config.hpp"
 #include "ssh/common/util.hpp"
+#include <algorithm>
 
 #ifdef HAVE_GETENTROPY
 #include <unistd.h>
@@ -64,7 +65,7 @@ struct crypto_context {
 	}
 
 	bool generate_random(securepath::ssh::span output) {
-		return handle && ::CryptGenRandom(handle, output.size(), securepath::ssh::to_uint8_ptr(output));
+		return handle && ::CryptGenRandom(handle, DWORD(output.size()), securepath::ssh::to_uint8_ptr(output));
 	}
 
 	HCRYPTPROV handle{};
